+++ packages/dmn-js-decision-table/test/spec/features/cell-selection/CellSelectionSpec.js
@@ -46,20 +46,20 @@ describe('features/cell-selection', function() {
       it('name', inject(function(cellSelection) {
 
         // when
-        click('decisionTable-name');
+        click('__decisionProperties_name');
 
         // then
-        expect(hasFocus('decisionTable-name')).to.be.true;
+        expect(hasFocus('__decisionProperties_name')).to.be.true;
       }));
 
 
       it('id', inject(function(cellSelection) {
 
         // when
-        click('decisionTable-id');
+        click('__decisionProperties_id');
 
         // then
-        expect(hasFocus('decisionTable-id')).to.be.true;
+        expect(hasFocus('__decisionProperties_id')).to.be.true;
       }));
 
     });
@@ -180,6 +180,21 @@ describe('features/cell-selection', function() {
         expectedSelection: 'outputEntry6'
       }));
 
+
+      it('in decision properties', inject(function(cellSelection) {
+
+        // given
+        click('__decisionProperties_id');
+
+        // when
+        const changed = cellSelection.selectCell('above');
+
+        // then
+        expect(hasFocus('__decisionProperties_name')).to.be.true;
+
+        expect(changed).to.be.true;
+      }));
+
     });
 
 
@@ -187,14 +202,14 @@ describe('features/cell-selection', function() {
 
       it('non-cell selection', inject(function(cellSelection) {
 
-        // when
-        click('decisionTable-id');
+        // given
+        click('__decisionProperties_name');
 
         // when
         const changed = cellSelection.selectCell('above');
 
         // then
-        expect(hasFocus('decisionTable-id')).to.be.true;
+        expect(hasFocus('__decisionProperties_name')).to.be.true;
 
         expect(changed).to.be.false;
       }));